change unsigned long long read_time() to unsigned long read_time()

+++ src/platform.cpp
@@ -92,7 +92,7 @@ static long long read_time(void)
 	return l;
 }
 #elif defined(__aarch64__)
-unsigned long long read_time()
+unsigned long read_time()
 {
 	unsigned long long val;
 	asm volatile ("mrs %0, cntvct_el0" : "=r" (val));
